\name{pmonitor}
\alias{pmonitor}
\title{plot the change of values across iterations}
\description{\code{plot} for monitoring.
}
\usage{
pmonitor(object, \dots)
}
\arguments{
 \item{object}{model object of class \code{"evolafit"}}
 \item{\dots}{Further arguments to be passed to the plot function.}
}
\value{
  vector of plot
 }
\author{Giovanny Covarrubias}
\seealso{\code{\link{plot}}, \code{\link{evolafit}}}
\keyword{models}

